+++ examples/app1/main.js
@@ -1,6 +1,8 @@
 var resolution = [1280, 720];
 var multisampling = 1;
 
+include("date.js");
+
 let black = new Color(0.0, 0.0, 0.0);
 let green = new Color(0.0, 1.0, 0.0, 0.8);
 let blue = new Color(0.0, 0.0, 1.0, 0.8);
@@ -17,28 +19,6 @@ let data;
 data = watch_json("text.json", (new_data) => {
   data = new_data;
 });
-
-function fmt_clock(date) {
-  let h = date.getHours();
-  let m = date.getMinutes();
-  let s = date.getSeconds();
-  let ampm = "am";
-  
-  if (h === 0) {
-    h = 12;
-  } else if (h === 12) {
-    ampm = "pm"
-  } else if (h > 12) {
-    h -= 12;
-    ampm = "pm";
-  }
-  
-  let h_str = h.toString().padStart(2);
-  let m_str = m.toString().padStart(2, "0");
-  let s_str = s.toString().padStart(2, "0");
-  
-  return `${h_str}:${m_str}:${s_str} ${ampm}`;
-}
 
 function init() { }
